refactor: Extract BarRight provider into BarRightWithProvider for mobile
