--- src/drive/web/modules/public/CozyBarRightMobile.jsx
@@ -1,17 +1,0 @@
-import { BarRight } from 'components/Bar'
-
-import React from 'react'
-
-import useBreakpoints from 'cozy-ui/transpiled/react/hooks/useBreakpoints'
-
-const CozyBarRightMobile = ({ children }) => {
-  const { isMobile } = useBreakpoints()
-
-  if (isMobile) {
-    return <BarRight>{children}</BarRight>
-  }
-
-  return <>{children}</>
-}
-
-export default CozyBarRightMobile
