+++ src/drive/web/modules/public/PublicToolbarCozyToCozy.jsx
@@ -1,20 +1,17 @@
 import React, { useState, useCallback } from 'react'
 
 import { useClient } from 'cozy-client'
-
-import BarContextProvider from 'cozy-ui/transpiled/react/BarContextProvider'
 import useBreakpoints from 'cozy-ui/transpiled/react/hooks/useBreakpoints'
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
 import Icon from 'cozy-ui/transpiled/react/Icon'
 import ActionMenu, { ActionMenuItem } from 'cozy-ui/transpiled/react/ActionMenu'
 
+import { BarRightWithProvider } from 'components/Bar'
 import { MoreButton } from 'components/Button'
 import SelectableItem from 'drive/web/modules/drive/Toolbar/selectable/SelectableItem'
 import { downloadFiles } from 'drive/web/modules/actions/utils'
-import CozyBarRightMobile from 'drive/web/modules/public/CozyBarRightMobile'
 import { DownloadFilesButton } from './DownloadButton'
 import { isFilesIsFile } from './PublicToolbarByLink'
-import { useWebviewIntent } from 'cozy-intent'
 
 const openExternalLink = url => (window.location = url)
 
@@ -74,33 +71,23 @@ const PublicToolbarCozyToCozy = ({
   files,
   isSharingShortcutCreated
 }) => {
-  const { t } = useI18n()
   const isFile = isFilesIsFile(files)
-  const client = useClient()
   const { isMobile } = useBreakpoints()
-  const webviewIntent = useWebviewIntent()
 
   const shouldDisplayMoreMenu = isMobile || (!isFile && files.length > 0)
 
   return (
-    <CozyBarRightMobile>
-      <BarContextProvider
-        client={client}
-        t={t}
-        store={client.store}
-        webviewService={webviewIntent}
-      >
-        {!isMobile && files.length > 0 && <DownloadFilesButton files={files} />}
-        {shouldDisplayMoreMenu && (
-          <MoreMenu
-            files={files}
-            discoveryLink={discoveryLink}
-            isSharingShortcutCreated={isSharingShortcutCreated}
-            isMobile={isMobile}
-          />
-        )}
-      </BarContextProvider>
-    </CozyBarRightMobile>
+    <BarRightWithProvider>
+      {!isMobile && files.length > 0 && <DownloadFilesButton files={files} />}
+      {shouldDisplayMoreMenu && (
+        <MoreMenu
+          files={files}
+          discoveryLink={discoveryLink}
+          isSharingShortcutCreated={isSharingShortcutCreated}
+          isMobile={isMobile}
+        />
+      )}
+    </BarRightWithProvider>
   )
 }
